refactor: 코드 리펙토링
- 클래스 이름 변경
