+++ lib/taste_registration/const/agent_speech_bubble_shape_type.dart
