+++ lib/taste_registration/const/interest.dart
