+++ lib/taste_registration/const/taste_state.dart
